refactor: reorganize operation schemas by domain structure
- Delete operationsSchema.ts and split into domain-based files
- Create schema/table.ts for table-related operations
- Move path validators to separate reusable utilities
- Update imports to use new schema structure
- Maintain existing functionality while improving organization
