+++ frontend/packages/db-structure/src/operation/index.ts
@@ -1,2 +1,2 @@
 export { applyPatchOperations } from './applyPatchOperations.js'
-export { operationsSchema } from './operationsSchema.js'
+export { operationsSchema } from './schema/index.js'
