+++ frontend/packages/db-structure/src/operation/pathValidators.ts
@@ -1,10 +1,6 @@
-import { PATH_PATTERNS } from './constants.js'
-
-const createPathValidator = (pattern: RegExp) => {
+export const createPathValidator = (pattern: RegExp) => {
   return (input: unknown): boolean => {
     if (typeof input !== 'string') return false
     return pattern.test(input)
   }
 }
-
-export const isTablePath = createPathValidator(PATH_PATTERNS.TABLE_BASE)
